requirements.txt
Pillow
gpuinfo
psutil
requests
